Rename the dialog components
