+++ src/js/components/PromptDialogComponent.jsx
@@ -1,15 +1,16 @@
 var React = require("react/addons");
 var classNames = require("classnames");
 
-var Util = require("../../helpers/Util");
-var ModalComponent = require("../ModalComponent");
+var Util = require("../helpers/Util");
+var ModalComponent = require("../components/ModalComponent");
 
-var AlertModalComponent = React.createClass({
-  displayName: "AlertModalComponent",
+var PromptModalComponent = React.createClass({
+  displayName: "PromptModalComponent",
 
   propTypes: {
     data: React.PropTypes.shape({
       actionButtonLabel: React.PropTypes.string.isRequired,
+      inputProperties: React.PropTypes.object.isRequired,
       message: React.PropTypes.string.isRequired,
       state: React.PropTypes.string.isRequired,
       title: React.PropTypes.string.isRequired
@@ -27,15 +28,23 @@ var AlertModalComponent = React.createClass({
   },
 
   componentDidMount: function () {
-    React.findDOMNode(this.refs.button).focus();
+    let input = React.findDOMNode(this.refs.input);
+    input.focus();
+    input.select();
   },
 
   handleAccept: function () {
-    this.props.onAccept();
+    this.props.onAccept(React.findDOMNode(this.refs.input).value);
   },
 
   handleDismiss: function () {
     this.props.onDismiss();
+  },
+
+  handleKeyUp: function (event) {
+    if (event.keyCode === 13) {
+      this.handleAccept();
+    }
   },
 
   render: function () {
@@ -53,15 +62,24 @@ var AlertModalComponent = React.createClass({
           {data.title}
         </div>
         <div className="modal-body">
-          {data.message}
+          <label>{data.message}</label>
+          <input className="form-control form-control-inverse"
+            {...data.inputProperties}
+            ref="input"
+            onKeyUp={this.handleKeyUp} />
         </div>
         <div className="modal-footer">
           <button
             className="btn btn-lg btn-success btn-inverse"
-            ref="button"
             type="button"
             onClick={this.handleAccept}>
             {data.actionButtonLabel}
+          </button>
+          <button
+            className="btn btn-lg btn-default btn-inverse"
+            type="button"
+            onClick={this.handleDismiss}>
+            Cancel
           </button>
         </div>
       </ModalComponent>
@@ -69,4 +87,4 @@ var AlertModalComponent = React.createClass({
   }
 });
 
-module.exports = AlertModalComponent;
+module.exports = PromptModalComponent;
